requirements.txt
aniso8601==3.0.2
Django==2.2
django-autofixture==0.12.1
django-filter==2.1.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
promise==2.2.1
psycopg2-binary==2.8.2
pytz==2019.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
django-graphql-jwt==0.2.1
